requirements.txt
appnope==0.1.0
arrow==0.10.0
awscli==1.11.170
awsebcli==3.12.0
beautifulsoup4==4.6.0
bleach==2.0.0
blessed==1.14.2
botocore==1.7.28
cement==2.8.2
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.7
decorator==4.1.2
Django==1.11.5
django-picklefield==1.0.0
django-q==0.8.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
future==0.16.0
get==0.0.0
gunicorn==19.7.1
html5lib==0.999999999
idna==2.6
ipykernel==4.6.1
ipython==6.2.0
ipython-genutils==0.2.0
ipywidgets==7.0.1
jedi==0.11.0
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
lxml==4.0.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.1.0
numpy==1.13.1
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.0
pathspec==0.5.0
pexpect==4.2.1
pickleshare==0.7.4
post==0.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
public==0.0.0
pyasn1==0.3.7
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
query-string==0.0.0
requests==2.9.1
rsa==3.4.2
s3transfer==0.1.11
semantic-version==2.5.0
setupfiles==0.0.50
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.14
tabulate==0.7.5
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.44.0
widgetsnbextension==3.0.3
